Improved tag creating ux

+++ routes/web.php
@@ -54,14 +54,14 @@ Route::middleware(AdminMiddleware::class)->group(function () {
     Route::get('/admin/users', [UserController::class, 'list'])->name('admin.users');
     Route::patch('/admin/users/{user}/promote', [UserController::class, 'promote'])->name('user.promote');
     Route::patch('/admin/users/{user}/demote', [UserController::class, 'demote'])->name('user.demote');
-    Route::get('/admin//user/create', [UserController::class, 'create'])->name('user.create');
+    Route::get('/admin/user/create', [UserController::class, 'create'])->name('user.create');
     Route::post('/admin/user/store', [UserController::class, 'store'])->name('user.store');
 
     Route::get('/admin/tags', [TagController::class, 'list'])->name('admin.tags');
-    Route::patch('/tags/{tag}/approve', [TagController::class, 'approve'])->name('tag.approve');
-    Route::delete('/tags/{tag}/delete', [TagController::class, 'destroy'])->name('tag.destroy');
-    Route::get('/tags/{tag}/edit', [TagController::class, 'edit'])->name('tag.edit');
-    Route::patch('/tags/{tag}/update', [TagController::class, 'update'])->name('tag.update');
+    Route::patch('/admin/tags/{tag}/approve', [TagController::class, 'approve'])->name('tag.approve');
+    Route::delete('/admin/tags/{tag}/delete', [TagController::class, 'destroy'])->name('tag.destroy');
+    Route::get('/admin/tags/{tag}/edit', [TagController::class, 'edit'])->name('tag.edit');
+    Route::patch('/admin/tags/{tag}/update', [TagController::class, 'update'])->name('tag.update');
 });
 
 Route::controller(UserController::class)->group(function () {
@@ -83,8 +83,8 @@ Route::controller(QuestionController::class)->group(function () {
     Route::get('/api/questions', 'fetch');
     Route::get('/api/questions/top', 'fetch');
     Route::get('/api/questions/followed', 'fetch');
-    Route::get('/api/questions/tag/{id}', 'fetch')->where('id', '[0-9]+');
-    Route::get('/api/questions/{id}/answers', 'fetch')->where('id', '[0-9]+');
+    Route::get('/api/questions/tag/{tag}', 'fetch')->where('tag', '[0-9]+');
+    Route::get('/api/questions/{question}/answers', 'fetch')->where('question', '[0-9]+');
 });
 
 Route::controller(AnswerController::class)->group(function () {
